+++ app/src/main/java/com/example/pokemonapipractice/presentation/ui/fragments/PokemonFragment.kt
@@ -1,4 +1,4 @@
-package com.example.pokemonapipractice.ui.fragments
+package com.example.pokemonapipractice.presentation.ui.fragments
 
 import android.util.Log
 import androidx.fragment.app.viewModels
@@ -7,9 +7,9 @@ import androidx.recyclerview.widget.LinearLayoutManager
 import by.kirich1409.viewbindingdelegate.viewBinding
 import com.example.pokemonapipractice.R
 import com.example.pokemonapipractice.databinding.FragmentPokemonBinding
-import com.example.pokemonapipractice.ui.adapters.PokemonAdapter
-import com.example.pokemonapipractice.ui.base.BaseFragment
-import com.example.pokemonapipractice.ui.state.UIState
+import com.example.pokemonapipractice.presentation.ui.adapters.PokemonAdapter
+import com.example.pokemonapipractice.presentation.ui.base.BaseFragment
+import com.example.pokemonapipractice.presentation.ui.state.UIState
 import dagger.hilt.android.AndroidEntryPoint
 import kotlinx.coroutines.launch
