--- app/src/main/java/com/example/pokemonapipractice/ui/base/BaseViewModel.kt
@@ -1,6 +1,0 @@
-package com.example.pokemonapipractice.ui.base
-
-import androidx.lifecycle.ViewModel
-
-open class BaseViewModel : ViewModel() {
-}
